+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 ## [v0.1.0](https://gitlab.com/mipimipi/yuppie/-/tags/v0.1.0) (2020-11-29)
 
 * Basic functionality for UPnP servers based on UPnP Device Architecture version 2.0
+
+## [v0.2.0](https://gitlab.com/mipimipi/yuppie/-/tags/v0.2.0) (???)
+
+### Changed
+
+* Refined error handling and error handling
+
+### Removed
+
+* Type `StateVars` removed. Use `map[string]StateVar` instead.